### 2020-02-01 #2059 @whedon
[PRE REVIEW]: fleck: Fast approximate light curves for starspot rotational modulation
**Submitting author:** @bmorris3 (<a href="http://orcid.org/0000-0003-2528-3409">Brett Morris</a>)
**Repository:** <a href="https://github.com/bmorris3/fleck" target ="_blank">https://github.com/bmorris3/fleck</a>
**Version:** v0.1
