### 2020-02-01 #2059 @dfm
Sure! I'm happy to edit. I'm mostly offline for the next few days, but I'll start the process before the end of the week.

@bmorris3: take a look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and let me know if you think anyone would be an especially good fit (but please don't @ them!).
