### 2020-02-05 #2059 @dfm
:wave: @rodluger, would be you available and willing to review this submission to JOSS? Thanks!
