### 2020-02-06 #2059 @dfm
@bmorris3: Thanks for letting us know. We will perform this review independently of AAS, so you should keep us posted on the status of that review. We will wait until both reviews are finished to publish the paper because the two will be linked at the end. Let me know if you have any questions!
