### 2020-02-13 #2059 @dfm
@rodluger: Thanks!

Here are the instructions: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

You'll see more in the review thread...
