### 2020-02-13 #2059 @dfm
@bmorris3, @rodluger, @nespinoza: The review is now started over in https://github.com/openjournals/joss-reviews/issues/2103. Let's head that way. Let me know if you have any questions!
