### 2020-02-13 #2103 @dfm
@rodluger, @nespinoza: Just a quick reminder that this has been submitted in collaboration with AAS. AAS makes a small donation to JOSS to support this process. For more information check out the following links and let me know if you have any issue with this:

- Blog post describing the collaboration: https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
- Summary of JOSS financial model: https://joss.theoj.org/about#costs
