### 2020-02-13 #2103 @nespinoza
Hi @bmorris3,

 First of all, thanks for this fantastic piece of software!

 I want to start my review by suggesting you mention in the documentation that this is a Python3-only package. I do know that this is now *the* Python standard, but it doesn't hurt to be nice to possible users --- especially given most OS older than ~last year come with Python 2.7 by default.

 In terms of the *functionality* of the code:

- I tried following the tutorial, but couldn't make the last plot of the "Getting started" tutorial (https://fleck.readthedocs.io/en/latest/fleck/gettingstarted.html). Just before the `plt.show()` call, when trying to make the `star.plot(` call, I get a: `'Star' object has no attribute 'plot'`. Perhaps this is a new function in the `Star` class that is on the Github but not in the pip-installable version of the code (which is how I installed `fleck`)?

 In terms of the *documentation* of the code:

- There is no explicit statement of need in the documentation. This is somewhat in the paper, but there is no introduction in the Read The Docs page for users that don't know what `fleck` is, what it does and what problems it is designed to do. I suggest creating an introductory page on the Read the Docs stating this.

- I think the community guidelines are not detailed enough, at least according to the checks we as reviewers have to do (see above). The Github repo does say that you "(...) love contributions", but it doesn't state explicitly how to submit contributions, if they must have, e.g., automated tests, etc. A simple sencence stating what you would want from contributions will suffice (e.g., "feel free to submit a pull request"). The same applies to any issues (I imagine you expect to use the Github issue tracker --- this should be stated) and support (e.g., should users reach out to you via e-mail? Github?).

 In terms of the *software paper*:

- I'm missing comparisons to other codes for rotational modulation and star-spotted transits such as e.g., `starry` (https://github.com/rodluger/starry), `spotrod` (https://github.com/bencebeky/spotrod) or `ellc` (https://github.com/pmaxted/ellc --- https://arxiv.org/pdf/1603.08484.pdf). Why is `fleck` needed in light of those existing algorithms? What is the advantage of using `fleck`?

- The paper mentions that the rapid computation of light curves allows to measure spot distributions with Approximate Bayesian Computation (ABC). Is there a reason why ABC is mentioned and not other bayesian optimization/posterior sampling methods? To be honest, that sentence makes me a little nervous because as a reviewer, I cannot know given the current state of the documentation if that statement is true or not. I suggest to either modify it to account for these concerns, remove it or add an optimization example in the documentation that uses either ABC or any other posterior sampling scheme out there.
