### 2020-02-14 70a6289
Adding more detailed description from paper to rtd index
