### 2020-02-14 760c1ba
Adding community guidelines
