### 2020-02-14 01bc6d7
Adding bayesian documentation
