+++ docs/fleck/bayes/mcmc.rst
@@ -1,0 +1,203 @@
+************
+MCMC Example
+************
+
+In this tutorial, we'll fit the light curve of the 23 ± 4 Myr-old transiting
+exoplanet star `V1298 Tau <https://arxiv.org/abs/1910.04563>`_ with ``fleck``.
+It is interesting to apply ``fleck`` to V1298 Tau because it is young and
+spotted, and has repeated spot signals over the duration of a few stellar
+rotations. Here, we'll use ``lightkurve`` to download the K2 photometry,
+``astropy`` to measure the stellar rotation period with a Lomb-Scargle,
+periodogram, and finally ``fleck`` to model the rotational modulation of the
+host star in order to invert the light curve for the spot properties: latitudes,
+longitudes, and radii.
+
+First, we import all of the required packages for this tutorial::
+
+    import matplotlib.pyplot as plt
+    import numpy as np
+    from astropy.coordinates import SkyCoord
+    from astropy.timeseries import LombScargle
+    import astropy.units as u
+    import healpy as hp
+    from lightkurve import search_lightcurvefile
+    from emcee import EnsembleSampler
+    from multiprocessing import Pool
+    from corner import corner
+    from fleck import Star
+
+Next we resolve the target coordinates using astropy, and download its light
+curve using ``lightkurve``'s handy method::
+
+    coord = SkyCoord.from_name('V1298 Tau')
+    slcf = search_lightcurvefile(coord, mission='K2')
+
+    lc = slcf.download_all()
+    pdcsap = lc.PDCSAP_FLUX.stitch()
+
+    time = pdcsap.time
+    flux = pdcsap.flux
+
+    notnanflux = ~np.isnan(flux)
+
+    flux = flux[notnanflux & (time > 2230) & (time < 2239)]
+    time = time[notnanflux & (time > 2230) & (time < 2239)]
+
+    flux /= np.mean(flux)
+
+Next, we want to determine the stellar rotation period, which will remain fixed
+in our ``fleck`` analysis. We do this with a Lomb-Scargle periodogram::
+
+    periods = np.linspace(1, 5, 1000) * u.day
+    freqs = 1 / periods
+    powers = LombScargle(time * u.day, flux).power(freqs)
+    best_period = periods[powers.argmax()]
+
+Now that we know that the best period is roughly 2.87 days, we can generate
+``fleck`` light curves that reproduce the spot modulation::
+
+    u_ld = [0.46, 0.11]
+    contrast = 0.7
+    phases = (time % best_period.value) / best_period.value
+    s = Star(contrast, u_ld, n_phases=len(time), rotation_period=best_period.value)
+
+    init_lons = np.array([0, 320, 100])
+    init_lats = [0, 20, 0]
+    init_rads = [0.01, 0.2, 0.3]
+
+    yerr = 0.002
+
+    init_p = np.concatenate([init_lons, init_lats, init_rads])
+
+With our initial parameters set, we now define the functions that ``emcee`` uses
+to do Markov Chain Monte Carlo::
+
+    def log_likelihood(p):
+        lons = p[0:3]
+        lats = p[3:6]
+        rads = p[6:9]
+
+        lc = s.light_curve(lons[:, None] * u.deg, lats[:, None] * u.deg, rads[:, None],
+                           inc_stellar=90*u.deg, times=time, time_ref=0)[:, 0]
+
+        return - 0.5 * np.sum((lc/np.mean(lc) - flux)**2 / yerr**2)
+
+    def log_prior(p):
+        lons = p[0:3]
[…]
