### 2020-02-14 #2103 @bmorris3
Thanks so much for the initial review, @nespinoza! I really appreciated your close eye.

RE Functionality:
* I _should_ have done a pip release prior to submitting the paper, sorry I did not! I just pushed a newer version to PyPI which you should be able to grab with `pip install --upgrade fleck`. I also intend to do another release and increment to v0.1 once the JOSS submission is accepted.

RE Documentation:
* You're right, I wrote most of the documentation while I was writing the code, that is to say, before I knew _exactly_ what I was going to do with it. I've updated the main landing page to reflect the text of the JOSS submission.
* I've added the required [statement about community contributions](https://github.com/bmorris3/fleck#contributing)

RE Software:
* The primary use case for fleck is *ensemble* generation of thousands of light curves for use with ABC methods. As far as I know, the packages you mentioned are good for modeling individual stellar systems, whereas fleck is optimized for vectorized, ensemble analysis. I think I describe this in the paper, but perhaps I could describe it better? Suggestions welcome!

* This is a great point – I was definitely letting the AAS Journals submission do the talking about the Bayesian methods and I failed to internally document how to use fleck for ABC. I've added docs examples with Bayesian techniques: one on [MCMC](https://fleck.readthedocs.io/en/latest/fleck/bayes/mcmc.html) and one on [ABC](https://fleck.readthedocs.io/en/latest/fleck/bayes/abc.html).
