### 2020-02-14 b24bd2a
Fixing docs typos
