### 2020-02-18 #2103 @nespinoza
Hi @bmorris3 --- thanks for the quick reply to the inquiries; I tested all the examples you added and all look great! I've checked the appropriate reviewer marks above, and all is almost ready to go on my side.

Regarding my comment on the paper, I think I understand that the methods and primary applications are different (i.e., you used this in the AAS submission to analyze _a sample_ of young stars, whereas the codes I mentioned above are used to model individual systems). However, in the documentation, you _do_ analyze individual stars (and spotted transits), and I believe this functionality and the caveats on that should be acknowledged in the paper or, at the very least, in the documentation itself. You could add something on the lines of "(...) while `fleck` is not intended to model individual stellar systems, it is able to do this reasonably well. However, other methods exist out there for this particular application such as X, Y, Z that might be better suited for those particular science cases".
