### 2020-02-19 247c34d
paper update to ref packages for individual LC modeling
