+++ docs/paper.bib
@@ -106,3 +106,42 @@ archivePrefix = "arXiv",
    adsurl = {http://adsabs.harvard.edu/abs/2011arXiv1102.1523V},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Luger:2019,
+       author = {{Luger}, Rodrigo and {Agol}, Eric and {Foreman-Mackey}, Daniel and
+         {Fleming}, David P. and {Lustig-Yaeger}, Jacob and {Deitrick}, Russell},
+        title = "{starry: Analytic Occultation Light Curves}",
+      journal = {\aj},
+     keywords = {eclipses, methods: analytical, occultations, techniques: photometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = "2019",
+        month = "Feb",
+       volume = {157},
+       number = {2},
+          eid = {64},
+        pages = {64},
+          doi = {10.3847/1538-3881/aae8e5},
+archivePrefix = {arXiv},
+       eprint = {1810.06559},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AJ....157...64L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Beky:2014,
+       author = {{B{\'e}ky}, Bence and {Kipping}, David M. and {Holman}, Matthew J.},
+        title = "{SPOTROD: a semi-analytic model for transits of spotted stars}",
+      journal = {\mnras},
+     keywords = {techniques: photometric, stars: individual: HAT-P-11- starspots, Astrophysics - Earth and Planetary Astrophysics},
+         year = "2014",
+        month = "Aug",
+       volume = {442},
+       number = {4},
+        pages = {3686-3699},
+          doi = {10.1093/mnras/stu1061},
+archivePrefix = {arXiv},
+       eprint = {1407.4465},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014MNRAS.442.3686B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
