### 2020-02-19 #2103 @bmorris3
Thanks for the reminder, sorry I didn't address that right away. I've updated the [text of the paper](https://github.com/bmorris3/fleck/blob/master/docs/paper.md#summary) to read:

> While fleck is intended to model ensembles of light curves, we also include some examples of fitting light curves of individual spotted stars. Other methods exist which are more efficient for modeling individual systems, such as starry [Luger:2019] and spotrod [Beky:2014], and we encourage users to explore whether those packages might be ideal for modeling individual systems.
