### 2020-02-19 #2103 @nespinoza
Thanks @bmorris3 for considering all the comments I made on this! @dfm, I think I have covered all the angles on my side of the review at least. On my side, I'm happy to recommend accepting `fleck` for publication pending the AAS submission status.
