### 2020-02-20 bmorris3/fleck#1 @rodluger
Sign error in spot longitude when plotting?
When I specify a spot longitude of 30 degrees, I appear to get a spot at a longitude of -30 degrees:

```python
from batman import TransitParams
import matplotlib.pyplot as plt
import numpy as np
import astropy.units as u
from fleck import Star

# Dummy planet
planet = TransitParams()
planet.a = 0.0
planet.inc = 0
planet.ecc = 0
planet.w = 0
planet.rp = 0

# Spot properties
inc_stellar = 90 * u.deg
spot_radii = np.array([[0.1]])
spot_lats = np.array([[0]]) * u.deg

# Longitude = 30 degrees
spot_lons = np.array([[30]]) * u.deg

# Show the star
star = Star(spot_contrast=0.7, u_ld=np.array([0, 0]), rotation_period=10)
star.plot(spot_lons, spot_lats, spot_radii, inc_stellar, time=0, planet=planet)
plt.show()
```

![image](https://user-images.githubusercontent.com/9323819/74949206-6cf9d400-53cb-11ea-92c0-7578985c4b2b.png)
