### 2020-02-20 bmorris3/fleck#2 @rodluger
Support `star.plot` when `planet=None`
Currently, if the user doesn't provide a planet instance when calling `Star.plot()`, several errors are raised, such as
```python
AttributeError: 'NoneType' object has no attribute 'a'
```
because the `planet` attribute is `None`. It might be useful to allow users interested in just visualizing the star to use the plotting functionality.

On a related note, some other optional keywords to [plot](https://fleck.readthedocs.io/en/latest/api/fleck.Star.html#fleck.Star.plot) default to `None` and lead to errors if left unspecified. These include the star's rotation period and the time, which should probably default to something like infinity and zero, respectively. Otherwise, I recommend you make them positional arguments.
