### 2020-02-20 bmorris3/fleck#3 @rodluger
Unphysical behavior when spot is on the limb
I think you might be [aware of this](https://fleck.readthedocs.io/en/latest/fleck/details.html#limitations-of-the-model), but the behavior is strange enough that it might be worth documenting it explicitly. When a spot is on the limb, a transiting planet causes the star to get brighter than its disk-integrated flux, which is unphysical:
```python
from batman import TransitParams
import matplotlib.pyplot as plt
import numpy as np
import astropy.units as u
from fleck import Star

planet = TransitParams()
planet.per = 88
planet.a = float(0.387 * u.AU / u.R_sun)
planet.rp = 0.1
planet.w = 90
planet.ecc = 0
planet.inc = 90
planet.t0 = 0
planet.limb_dark = "quadratic"
planet.u = [0, 0]
planet.lam = 0

inc_stellar = 90 * u.deg
spot_radii = np.array([[0.5]])
spot_lons = np.array([[80]]) * u.deg
spot_lats = np.array([[0]]) * u.deg

times = np.linspace(-0.5, 0.5, 5000)

fig, ax = plt.subplots(1, 2, figsize=(8, 2.5))
star = Star(spot_contrast=0.7, u_ld=planet.u, rotation_period=np.inf)
star.plot(
    spot_lons, spot_lats, spot_radii, inc_stellar, planet=planet, time=0, ax=ax[0]
)
ax[0].set(xlabel="$\hat{X}$", ylabel="$\hat{Y}$")
lc = star.light_curve(
    spot_lons, spot_lats, spot_radii, inc_stellar, planet=planet, times=times
)
ax[1].plot(times, lc, color="k")
ax[1].set(xlabel="Time [d]", ylabel="Flux")
fig.tight_layout()
plt.show()
```

![image](https://user-images.githubusercontent.com/9323819/74951113-e1357700-53cd-11ea-8bcc-9ea6c1a59125.png)

As you mention in the docs, this occurs mostly for large spots, and I think it's due to the fact that your spots are 2 dimensional and don't curve along the surface of the sphere.
