### 2020-02-20 bmorris3/fleck#3 @bmorris3
Yep I'm aware of this. That's why the paper and docs mostly focus on how to use fleck for rotational modulation. It's definitely not meant to stand-in for models like `starry`.
