### 2020-02-25 #2103 @rodluger
Hi, sorry for the delay. I'll have this done by tomorrow. In the meantime, I raised three issues ([here](https://github.com/bmorris3/fleck/issues/1), [here](https://github.com/bmorris3/fleck/issues/2), and [here](https://github.com/bmorris3/fleck/issues/3)). I believe the first one is a bug in the plotting routine, the second is a user interface suggestion, and the third is a limitation of the model that has already been addressed & closed. @bmorris3, can you take a look at the remaining two issues when you get a chance? I'll have final comments in the next 24 hours.
