### 2020-02-25 #2103 @rodluger
@dfm, is there a way to re-generate the article proof after Brett's recent changes?
