### 2020-02-26 #2103 @dfm
@rodluger: Thanks!

@bmorris3: Can you report back here when:

1. You have address the final issues that @rodluger raised, and
2. You have the DOI of the accepted AAS paper. For reference, the DOI of this paper (when it is published) will be 10.21105.joss.02103. I believe that the AAS folks will want that.
