### 2020-03-01 b6466e0
More flexible plotting function
