### 2020-03-01 bmorris3/fleck#2 @bmorris3
Thanks for this bug report. I've fixed this up in b6466e03730cc68fa966b2be2ba94b1ae7ff1078, you can now do the following:

```python
import matplotlib.pyplot as plt
from fleck import Star
import astropy.units as u
import numpy as np

spot_contrast = 0.7
u_ld = [0.5079, 0.2239]

n_phases = 30
lons = [[60]] * u.deg
lats = [[10]] * u.deg
radii = np.array([[0.1]])
inc_stellar = 90 * u.deg
stars = Star(spot_contrast=spot_contrast, n_phases=n_phases, u_ld=u_ld,
             rotation_period=10)
stars.plot(lons, lats, radii, inc_stellar, time=0)
plt.show()
```

![example1](https://user-images.githubusercontent.com/3497584/75627994-034ca780-5bcd-11ea-8bcb-cf0aa62fbd82.png)
