### 2020-03-01 bmorris3/fleck#1 @bmorris3
Thanks for the catch! There was a double sign error happening in the code which I've now corrected in 1b5eef1.
