### 2020-03-02 #2103 @dfm
Last thing (I think) for now: can you update the BibTeX entry for Numpy to include the DOI (https://doi.org/10.1109/mcse.2011.37)?

Then I'll pause this review until you have the DOI from the AAS journals.
