### 2020-03-03 dd8993c
adding DOI for numpy [ci skip]
