+++ docs/paper.bib
@@ -105,6 +105,7 @@ archivePrefix = "arXiv",
     month = feb,
    adsurl = {http://adsabs.harvard.edu/abs/2011arXiv1102.1523V},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      doi = {10.1109/mcse.2011.37}
 }
 
 @ARTICLE{Luger:2019,
