### 2020-03-03 #2103 @bmorris3
Ok, I think I've [done it](https://github.com/bmorris3/fleck/blob/master/docs/paper.bib#L108) correctly.
