+++ docs/paper.bib
@@ -104,7 +104,7 @@ archivePrefix = "arXiv",
      year = 2011,
     month = feb,
    adsurl = {http://adsabs.harvard.edu/abs/2011arXiv1102.1523V},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System},
       doi = {10.1109/mcse.2011.37}
 }
 
