### 2020-03-04 #2103 @dfm
@bmorris3: I submitted one more pull request with a BibTeX fix. After merging that, we should be good to go. I'll pause this review until the AAS paper has a DOI (when it is published). When that happens:

1. Update the `paper.md` file with the AAS journal name and DOI,
2. Create a new versioned release in the archive (make sure that the title and author list match the JOSS submission), and
3. Report back here with the new Zenodo (or similar) DOI and version number.
