### 2020-03-06 #2103 @bmorris3
@dfm – `paper.md` has been updated, Zenodo DOI minted (https://doi.org/10.5281/zenodo.3700053) and [version number 1.0](https://github.com/bmorris3/fleck/releases/tag/v1.0) released.
