### 2020-03-06 #2103 @Kevin-Mattheus-Moerman
Please check the following:
- [ ] Please make sure the Zenodo meta data such as the title and the author list match the JOSS paper. You can just edit this information on Zenodo. Currently both the title and the author set are different.

- [ ] The statement concerning "...in about 10 seconds on a 2.5 GHz Intel Core i7 processor." is a bit vague. If you want to cite something like this I suggest you add more detail. There are various CPU's with a 2.5 GHz base frequency with varying degrees of performance. For instance the [i7-2860QM](https://ark.intel.com/content/www/us/en/ark/products/53476/intel-core-i7-2860qm-processor-8m-cache-up-to-3-60-ghz.html) has 4 cores, 8 threads and a 8Mb Cache, while the [i7-6500U](https://ark.intel.com/content/www/us/en/ark/products/88194/intel-core-i7-6500u-processor-4m-cache-up-to-3-10-ghz.html) only has 2 cores, 4 threads, and a 4Mb Cache. Also, computational time could highly depend on the system RAM. So perhaps say something like:
"...in about 10 seconds on a regular personal computer (64-bit [Insert OS here], Intel Core i7 XXXXXX CPU, [insert memory amount] Gb RAM)." (I was assuming you are referring to performance on a standard desktop/laptop but change as you see fit).
