### 2020-03-09 #2103 @bmorris3
Hi @Kevin-Mattheus-Moerman – sorry about the mismatch, do I create and edit a `.zenodo.json` file to edit the repo's metadata in the existing release?
