### 2020-03-09 #2103 @bmorris3
Hi @Kevin-Mattheus-Moerman, I've edited the `paper.md` file according to your recommendations and fixed the zenodo metadata. Thanks!
