### 2020-03-09 #2103 @Kevin-Mattheus-Moerman
@bmorris3 thanks the Zenodo archive looks good. About the CPU specification. It still only mentions the base frequency. Can you add either the CPU id (I'm guessing you have the [i7-4870HQ](https://ark.intel.com/content/www/us/en/ark/products/83504/intel-core-i7-4870hq-processor-6m-cache-up-to-3-70-ghz.html), but check which one you have) or quote the number of cores, number of threads, and the cache size?
