### 2020-03-09 #2103 @bmorris3
Sorry for my lack of precision, I had to dig a bit to verify the processor specs. [Done!](https://github.com/bmorris3/fleck/blob/master/docs/paper.md)
